Added config option for CommonJS syntax
This is controversial. Currently, passing a configuration by using
`rangy.config.alertOnFail = false` after `require('rangy')` does not work
in all environments, I believe because it is async and it isn't able to
access the function as it is wrapped in an IIFE. In order to fix this,
when rangy is initialized, the config has to be loaded in immediately.

This isn't possible with the current syntax, as you can't load in parameters.
After a lot of fiddling around, the best I could come up with was this
solution. However, it is a breaking change - all implementations using require
and CommonJS will have to use this syntax:

```js
var rangy = require('rangy')()
```

To set a config, you would do this:

```js
var rangy = require('rangy')({'alertOnFail': true})
```

I couldn't find a way around this. This solution works for me, closes #280.